# docs/swagger.yaml
definitions:
...
  models.Notification:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      is_read:
        type: boolean
      message:
        type: string
      related_id:
        description: ID of related load, trip, etc.
        type: integer
      title:
        type: string
      type:
        description: QUOTE_RECEIVED, LOAD_BOOKED, PICKUP_SCHEDULED, etc.
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Quote:
    properties:
      accepted_at:
        type: string
      carrier_id:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      deleted_at:
        type: string
      delivery_date:
        type: string
      id:
        type: integer
      load_id:
        type: integer
      notes:
        type: string
      pickup_date:
        type: string
      quote_amount:
        type: number
      status:
        description: PENDING, ACCEPTED, REJECTED, EXPIRED
        type: string
      updated_at:
        type: string
      valid_until:
        type: string
    type: object
...
info:
  contact: {}
  title: Triplink API
  version: "1.0"
paths:
...
  /carriers/{carrier_id}/quotes:
    get:
      description: Get all quotes created by a specific carrier
      parameters:
      - description: Carrier ID
        in: path
        name: carrier_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Quote'
            type: array
      summary: Get quotes by carrier
      tags:
      - quotes
...
  /loads/{load_id}/quotes:
    get:
      description: Get all quotes for a specific load
      parameters:
      - description: Load ID
        in: path
        name: load_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Quote'
            type: array
      summary: Get quotes for a load
      tags:
      - quotes
...
  /notifications:
    post:
      consumes:
      - application/json
      description: Create a new notification for a user
      parameters:
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/models.Notification'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Notification'
      summary: Create a notification
      tags:
      - notifications
  /notifications/{id}:
    delete:
      description: Delete a specific notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete notification
      tags:
      - notifications
  /notifications/{id}/read:
    put:
      description: Mark a specific notification as read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Notification'
      summary: Mark notification as read
      tags:
      - notifications
  /quotes:
    post:
      consumes:
      - application/json
      description: Create a new quote for a load
      parameters:
      - description: Quote data
        in: body
        name: quote
        required: true
        schema:
          $ref: '#/definitions/models.Quote'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Quote'
      summary: Create a quote
      tags:
      - quotes
  /quotes/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing quote (only if pending)
      parameters:
      - description: Quote ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated quote data
        in: body
        name: quote
        required: true
        schema:
          $ref: '#/definitions/models.Quote'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Quote'
      summary: Update quote
      tags:
      - quotes
  /quotes/{id}/accept:
    post:
      description: Accept a quote and create a booking
      parameters:
      - description: Quote ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Quote'
      summary: Accept a quote
      tags:
      - quotes
  /quotes/{id}/reject:
    post:
      description: Reject a quote
      parameters:
      - description: Quote ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Quote'
      summary: Reject a quote
      tags:
      - quotes
...
  /users/{user_id}/notifications:
    get:
      description: Get all notifications for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Show only unread notifications
        in: query
        name: unread_only
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Notification'
            type: array
      summary: Get user notifications
      tags:
      - notifications
  /users/{user_id}/notifications/count:
    get:
      description: Get notification counts for a user (total and unread)
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get notification counts
      tags:
      - notifications
  /users/{user_id}/notifications/read-all:
    put:
      description: Mark all notifications for a user as read
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Mark all notifications as read
      tags:
      - notifications
...
swagger: "2.0"